uniformize ups extension implementation function names
