+++ extensions/ups/purplship/providers/ups/__init__.py
@@ -1,3 +1,9 @@
 from purplship.providers.ups.utils import Settings
-from purplship.providers.ups.track import track_request, parse_track_response
-from purplship.providers.ups.address_validation import address_validation_request, parse_address_validation_response
+from purplship.providers.ups.address import (
+    parse_address_validation_response,
+    address_validation_request
+)
+from purplship.providers.ups.tracking import (
+    parse_tracking_response,
+    tracking_request,
+)
