--- extensions/ups/purplship/providers/ups/freight/__init__.py
@@ -1,5 +1,0 @@
-from purplship.providers.ups.freight.ship import (
-    shipment_request,
-    parse_shipment_response,
-)
-from purplship.providers.ups.freight.rate import rate_request, parse_rate_response
