--- extensions/ups/purplship/providers/ups/package/ship/__init__.py
@@ -1,8 +1,0 @@
-from purplship.providers.ups.package.ship.create import (
-    shipment_request,
-    parse_shipment_response,
-)
-from purplship.providers.ups.package.ship.void import (
-    void_shipment_request,
-    parse_void_shipment_response,
-)
