+++ extensions/ups/purplship/providers/ups/package/shipment/cancel.py
@@ -15,7 +15,7 @@ from purplship.providers.ups.utils import Settings, default_request_serializer
 from purplship.providers.ups.error import parse_error_response
 
 
-def parse_void_shipment_response(
+def parse_shipment_cancel_response(
     response: Element, settings: Settings
 ) -> Tuple[ConfirmationDetails, List[Message]]:
     status = build(
@@ -40,7 +40,7 @@ def parse_void_shipment_response(
     return cancellation, parse_error_response(response, settings)
 
 
-def void_shipment_request(
+def shipment_cancel_request(
     payload: ShipmentCancelRequest, settings: Settings
 ) -> Serializable[Envelope]:
