--- extensions/ups_freight/purplship/mappers/ups_freight/__init__.py
@@ -1,3 +1,0 @@
-from purplship.mappers.ups_freight.mapper import Mapper
-from purplship.mappers.ups_freight.proxy import Proxy
-from purplship.mappers.ups_freight.settings import Settings
