--- extensions/ups_freight/purplship/mappers/ups_freight/settings.py
@@ -1,21 +1,0 @@
-"""Purplship UPS connection settings."""
-
-import attr
-from purplship.providers.ups.utils import Settings as BaseSettings
-
-
-@attr.s(auto_attribs=True)
-class Settings(BaseSettings):
-    """UPS connection settings."""
-
-    username: str
-    password: str
-    access_license_number: str
-    account_number: str = None
-    id: str = None
-    test: bool = False
-    carrier_id: str = "ups_freight"
-
-    @property
-    def carrier_name(self):
-        return "ups_freight"
