+++ extensions/ups_package/purplship/mappers/ups_package/mapper.py
@@ -1,46 +1,47 @@
 from typing import List, Tuple
-from pyups.av_request import AddressValidationRequest as UPSAddressValidationRequest
+from purplship.core.utils.serializable import Serializable, Deserializable
 from purplship.api.mapper import Mapper as BaseMapper
-from purplship.mappers.ups_package.settings import Settings
-from purplship.core.utils import Deserializable, Serializable, Envelope, Pipeline
 from purplship.core.models import (
-    RateRequest,
-    ShipmentRequest,
-    TrackingRequest,
-    ShipmentDetails,
-    RateDetails,
-    TrackingDetails,
-    Message,
-    PickupRequest,
+    AddressValidationRequest,
+    ShipmentCancelRequest,
     PickupUpdateRequest,
     PickupCancelRequest,
-    PickupDetails,
-    ConfirmationDetails,
-    AddressValidationRequest,
+    ShipmentRequest,
+    TrackingRequest,
+    PickupRequest,
+    RateRequest,
+
     AddressValidationDetails,
-    ShipmentCancelRequest,
-)
-from purplship.providers.ups import (
-    parse_track_response,
-    track_request,
-    address_validation_request,
-    parse_address_validation_response,
+    ConfirmationDetails,
+    TrackingDetails,
+    ShipmentDetails,
+    PickupDetails,
+    RateDetails,
+    Message,
 )
 from purplship.providers.ups.package import (
+    parse_shipment_cancel_response,
+    parse_pickup_update_response,
+    parse_pickup_cancel_response,
     parse_shipment_response,
-    parse_rate_response,
-    shipment_request,
-    rate_request,
-    create_pickup_pipeline,
     parse_pickup_response,
-    update_pickup_pipeline,
-    cancel_pickup_request,
-    parse_cancel_pickup_response,
-    void_shipment_request,
-    parse_void_shipment_response,
+    parse_rate_response,
+
+    shipment_cancel_request,
+    pickup_update_request,
+    pickup_cancel_request,
+    shipment_request,
+    pickup_request,
+    rate_request,
 )
-from pyups.track_web_service_schema import TrackRequest
-from pyups.ship_web_service_schema import ShipmentRequest as UPSShipmentRequest
+from purplship.providers.ups import (
+    parse_address_validation_response,
+    parse_tracking_response,
+
+    address_validation_request,
+    tracking_request,
+)
+from purplship.mappers.fedex_express.settings import Settings
 
 
 class Mapper(BaseMapper):
@@ -48,37 +49,41 @@ class Mapper(BaseMapper):
 
     """Request Mappers"""
 
-    def create_address_validation_request(self, payload: AddressValidationRequest) -> Serializable[UPSAddressValidationRequest]:
+    def create_address_validation_request(self, payload: AddressValidationRequest) -> Serializable:
         return address_validation_request(payload, self.settings)
 
-    def create_rate_request(self, payload: RateRequest) -> Serializable[Envelope]:
+    def create_rate_request(
+        self, payload: RateRequest
+    ) -> Serializable:
         return rate_request(payload, self.settings)
 
     def create_tracking_request(
         self, payload: TrackingRequest
-    ) -> Serializable[List[TrackRequest]]:
-        return track_request(payload, self.settings)
+    ) -> Serializable:
+        return tracking_request(payload, self.settings)
 
     def create_shipment_request(
         self, payload: ShipmentRequest
-    ) -> Serializable[UPSShipmentRequest]:
+    ) -> Serializable:
         return shipment_request(payload, self.settings)
 
-    def create_cancel_shipment_request(self, payload: ShipmentCancelRequest) -> Serializable:
-        return void_shipment_request(payload, self.settings)
-
-    def create_pickup_request(self, payload: PickupRequest) -> Serializable[Pipeline]:
-        return create_pickup_pipeline(payload, self.settings)
+    def create_pickup_request(
+        self, payload: PickupRequest
+    ) -> Serializable:
+        return pickup_request(payload, self.settings)
 
     def create_pickup_update_request(
         self, payload: PickupUpdateRequest
-    ) -> Serializable[Pipeline]:
-        return update_pickup_pipeline(payload, self.settings)
+    ) -> Serializable:
+        return pickup_update_request(payload, self.settings)
 
     def create_cancel_pickup_request(
         self, payload: PickupCancelRequest
-    ) -> Serializable[Envelope]:
-        return cancel_pickup_request(payload, self.settings)
+    ) -> Serializable:
+        return pickup_cancel_request(payload, self.settings)
+
+    def create_cancel_shipment_request(self, payload: ShipmentCancelRequest) -> Serializable[str]:
+        return shipment_cancel_request(payload, self.settings)
 
     """Response Parsers"""
 
@@ -87,37 +92,37 @@ class Mapper(BaseMapper):
     ) -> Tuple[AddressValidationDetails, List[Message]]:
         return parse_address_validation_response(response.deserialize(), self.settings)
 
+    def parse_cancel_pickup_response(
+        self, response: Deserializable[str]
+    ) -> Tuple[ConfirmationDetails, List[Message]]:
+        return parse_pickup_cancel_response(response.deserialize(), self.settings)
+
+    def parse_cancel_shipment_response(
+        self, response: Deserializable
+    ) -> Tuple[ConfirmationDetails, List[Message]]:
+        return parse_shipment_cancel_response(response.deserialize(), self.settings)
+
+    def parse_pickup_response(
+        self, response: Deserializable[str]
+    ) -> Tuple[PickupDetails, List[Message]]:
+        return parse_pickup_response(response.deserialize(), self.settings)
+
+    def parse_pickup_update_response(
+        self, response: Deserializable[str]
+    ) -> Tuple[PickupDetails, List[Message]]:
+        return parse_pickup_update_response(response.deserialize(), self.settings)
+
     def parse_rate_response(
         self, response: Deserializable[str]
     ) -> Tuple[List[RateDetails], List[Message]]:
         return parse_rate_response(response.deserialize(), self.settings)
-
-    def parse_tracking_response(
-        self, response: Deserializable[str]
-    ) -> Tuple[List[TrackingDetails], List[Message]]:
-        return parse_track_response(response.deserialize(), self.settings)
 
     def parse_shipment_response(
         self, response: Deserializable[str]
     ) -> Tuple[ShipmentDetails, List[Message]]:
         return parse_shipment_response(response.deserialize(), self.settings)
 
-    def parse_cancel_shipment_response(
+    def parse_tracking_response(
         self, response: Deserializable[str]
-    ) -> Tuple[ConfirmationDetails, List[Message]]:
-        return parse_void_shipment_response(response.deserialize(), self.settings)
-
-    def parse_pickup_response(
-        self, response: Deserializable
-    ) -> Tuple[PickupDetails, List[Message]]:
-        return parse_pickup_response(response.deserialize(), self.settings)
-
-    def parse_pickup_update_response(
-        self, response: Deserializable
-    ) -> Tuple[PickupDetails, List[Message]]:
-        return parse_pickup_response(response.deserialize(), self.settings)
-
-    def parse_cancel_pickup_response(
-        self, response: Deserializable
-    ) -> Tuple[ConfirmationDetails, List[Message]]:
-        return parse_cancel_pickup_response(response.deserialize(), self.settings)
+    ) -> Tuple[List[TrackingDetails], List[Message]]:
+        return parse_tracking_response(response.deserialize(), self.settings)
