+++ tests/dhl_express/shipment.py
@@ -229,7 +229,7 @@ ShipmentRequestXml = f"""<req:ShipmentRequest xmlns:req="http://www.dhl.com" xml
         <City>Brussels</City>
         <PostalCode>1060</PostalCode>
         <CountryCode>BE</CountryCode>
-        <CountryName>BELGIUM</CountryName>
+        <CountryName>Belgium</CountryName>
         <Contact>
             <PersonName>Mrs Orlander</PersonName>
             <PhoneNumber>506-851-2271</PhoneNumber>
@@ -281,7 +281,7 @@ ShipmentRequestXml = f"""<req:ShipmentRequest xmlns:req="http://www.dhl.com" xml
         <DivisionCode>AZ</DivisionCode>
         <PostalCode>85260</PostalCode>
         <CountryCode>US</CountryCode>
-        <CountryName>UNITED STATES OF AMERICA</CountryName>
+        <CountryName>United States</CountryName>
         <Contact>
             <PersonName>Ms Lucian</PersonName>
             <PhoneNumber>1 23 8613402</PhoneNumber>
